--- nuplan/planning/simulation/planner/pdm_planner/abstract_pdm_planner.py
@@ -1,252 +1,0 @@
-from abc import ABC
-from typing import Dict, List, Optional, Tuple
-
-import numpy as np
-import numpy.typing as npt
-from shapely.geometry import Point
-
-from nuplan.common.actor_state.ego_state import EgoState
-from nuplan.common.actor_state.state_representation import StateSE2
-from nuplan.common.actor_state.vehicle_parameters import VehicleParameters
-
-from nuplan.common.maps.abstract_map import AbstractMap
-from nuplan.common.maps.abstract_map_objects import (
-    LaneGraphEdgeMapObject,
-    RoadBlockGraphEdgeMapObject,
-)
-from nuplan.common.maps.maps_datatypes import SemanticMapLayer
-from nuplan.planning.simulation.planner.abstract_planner import AbstractPlanner
-from nuplan.planning.simulation.planner.pdm_planner.observation.pdm_observation import (
-    PDMObservation,
-)
-from nuplan.planning.simulation.planner.pdm_planner.proposal.batch_idm_policy import BatchIDMPolicy
-from nuplan.planning.simulation.planner.pdm_planner.proposal.pdm_generator import PDMGenerator
-from nuplan.planning.simulation.planner.pdm_planner.proposal.pdm_proposal import (
-    PDMProposalManager,
-)
-from nuplan.planning.simulation.planner.pdm_planner.utils.pdm_path import (
-    PDMPath,
-)
-from nuplan.planning.simulation.planner.pdm_planner.scoring.pdm_scorer import PDMScorer
-from nuplan.planning.simulation.planner.pdm_planner.simulation.pdm_simulator import PDMSimulator
-from nuplan.planning.simulation.planner.pdm_planner.utils.pdm_geometry_utils import (
-    normalize_angle,
-    parallel_discrete_path,
-)
-from nuplan.planning.simulation.planner.pdm_planner.utils.graph_search.dijkstra import Dijkstra
-from nuplan.planning.simulation.planner.pdm_planner.utils.pdm_emergency_brake import (
-    PDMEmergencyBrake,
-)
-
-
-class AbstractPDMPlanner(AbstractPlanner, ABC):
-    """
-    Interface for PDM planners. Inherit from this class to use any PDM planner variant (Closed, Hybrid, etc.)
-    """
-
-    def __init__(
-        self,
-        idm_policies: BatchIDMPolicy,
-        lateral_offsets: Optional[List[float]],
-        trajectory_samples: int,
-        proposal_samples: int,
-        sample_interval: float,
-        map_radius: float,
-    ):
-        """
-        Constructor for AbstractPDMPlanner
-        :param idm_policies: BatchIDMPolicy class
-        :param lateral_offsets: centerline offsets for proposals (optional)
-        :param trajectory_samples: number of trajectory samples
-        :param proposal_samples: number of proposal samples
-        :param sample_interval: interval of trajectory/proposal samples
-        :param map_radius: radius around ego to consider
-        """
-
-        # config parameters
-        self._idm_policies: BatchIDMPolicy = idm_policies
-        self._lateral_offsets: Optional[List[float]] = lateral_offsets
-        self._trajectory_samples: int = trajectory_samples
-        self._proposal_samples: int = proposal_samples
-        self._sample_interval: float = sample_interval  # [s]
-        self._map_radius: int = map_radius  # [m]
-
-        # observation/forecasting class
-        self._observation = PDMObservation(trajectory_samples, proposal_samples, sample_interval)
-
-        # proposal/trajectory related classes
-        self._generator = PDMGenerator(trajectory_samples, proposal_samples, sample_interval)
-        self._simulator = PDMSimulator(proposal_samples, sample_interval)
-        self._scorer = PDMScorer(proposal_samples, sample_interval)
-        self._emergency_brake = PDMEmergencyBrake(trajectory_samples, sample_interval)
-
-        # lazy loaded
-        self._map_api: Optional[AbstractMap] = None
-        self._route_roadblock_dict: Optional[Dict[str, RoadBlockGraphEdgeMapObject]] = None
-        self._route_lane_dict: Optional[Dict[str, LaneGraphEdgeMapObject]] = None
-
-        self._centerline: Optional[PDMPath] = None
-        self._vehicle_parameters: Optional[VehicleParameters] = None
-
-        self._proposal_manager: Optional[PDMProposalManager] = None
-        self._iteration: int = 0
-
-    def _update_proposal_manager(self, ego_state: EgoState):
-        """
-        Updates or initializes PDMProposalManager class
-        :param ego_state: state of ego-vehicle
-        """
-
-        current_lane = self._get_starting_lane(ego_state)
-
-        # TODO: Find additional conditions to trigger re-planning
-        create_new_proposals = self._iteration == 0
-
-        if create_new_proposals:
-            proposal_paths: List[PDMPath] = self._get_proposal_paths(current_lane)
-            self._centerline = proposal_paths[0]
-
-            self._proposal_manager = PDMProposalManager(
-                lateral_proposals=proposal_paths,
-                longitudinal_policies=self._idm_policies,
-            )
-
-        # update proposals
-        self._proposal_manager.update(current_lane.speed_limit_mps)
-
-    def _get_proposal_paths(self, current_lane: LaneGraphEdgeMapObject) -> List[PDMPath]:
-        """
-        Returns a list of path's to follow for the proposals. Inits a centerline.
-        :param current_lane: current or starting lane of path-planning
-        :return: lists of paths (0-index is centerline)
-        """
-        centerline_discrete_path = self._get_discrete_centerline(current_lane)
-
-        # 1. save centerline path (necessary for progress metric)
-        self._centerline = PDMPath(centerline_discrete_path)
-        output_paths: List[PDMPath] = [self._centerline]
-
-        # 2. add additional paths with lateral offset of centerline
-        if self._lateral_offsets is not None:
-            for lateral_offset in self._lateral_offsets:
-                offset_discrete_path = parallel_discrete_path(
-                    discrete_path=centerline_discrete_path, offset=lateral_offset
-                )
-                output_paths.append(PDMPath(offset_discrete_path))
-
-        return output_paths
-
-    def _load_route_dicts(self, route_roadblock_ids: List[str]) -> None:
-        """
-        Loads roadblock and lane dictionaries of the target route from the map-api.
-        :param route_roadblock_ids: ID's of on-route roadblocks
-        """
-
-        # remove repeated ids while remaining order in list
-        route_roadblock_ids = list(dict.fromkeys(route_roadblock_ids))
-
-        self._route_roadblock_dict = {}
-        self._route_lane_dict = {}
-
-        for id_ in route_roadblock_ids:
-            block = self._map_api.get_map_object(id_, SemanticMapLayer.ROADBLOCK)
-            block = block or self._map_api.get_map_object(id_, SemanticMapLayer.ROADBLOCK_CONNECTOR)
-
-            self._route_roadblock_dict[block.id] = block
-
-            for lane in block.interior_edges:
-                self._route_lane_dict[lane.id] = lane
-
-    def _get_discrete_centerline(
-        self, current_lane: LaneGraphEdgeMapObject, search_depth: int = 30
-    ) -> List[StateSE2]:
-        """
-        Applies a Dijkstra search on the lane-graph to retrieve discrete centerline.
-        :param current_lane: _description_
-        :param search_depth: depth of search (for runtime), defaults to 30
-        :return: list of discrete states on centerline (x,y,θ)
-        """
-
-        roadblocks = list(self._route_roadblock_dict.values())
-        roadblock_ids = list(self._route_roadblock_dict.keys())
-
-        # find current roadblock index
-        start_idx = np.argmax(np.array(roadblock_ids) == current_lane.get_roadblock_id())
-        roadblock_window = roadblocks[start_idx : start_idx + search_depth]
-
-        graph_search = Dijkstra(current_lane, list(self._route_lane_dict.keys()))
-        route_plan, path_found = graph_search.search(roadblock_window[-1])
-
-        centerline_discrete_path: List[StateSE2] = []
-        for lane in route_plan:
-            centerline_discrete_path.extend(lane.baseline_path.discrete_path)
-
-        return centerline_discrete_path
-
-    def _get_starting_lane(self, ego_state: EgoState) -> LaneGraphEdgeMapObject:
-        """
-        Returns the most suitable starting lane, in ego's vicinity.
-        :param ego_state: state of ego-vehicle
-        :return: lane object (on-route)
-        """
-        starting_lane: LaneGraphEdgeMapObject = None
-        on_route_lanes, heading_error = self._get_intersecting_lanes(ego_state)
-
-        if on_route_lanes:
-            # 1. Option: find lanes from lane occupancy-map
-            # select lane with lowest heading error
-            starting_lane = on_route_lanes[np.argmin(np.abs(heading_error))]
-            return starting_lane
-
-        else:
-            # 2. Option: find any intersecting or close lane on-route
-            closest_distance = np.inf
-            for edge in self._route_lane_dict.values():
-                if edge.contains_point(ego_state.center):
-                    starting_lane = edge
-                    break
-
-                distance = edge.polygon.distance(ego_state.car_footprint.geometry)
-                if distance < closest_distance:
-                    starting_lane = edge
-                    closest_distance = distance
-
-        return starting_lane
-
-    def _get_intersecting_lanes(
-        self, ego_state: EgoState
-    ) -> Tuple[List[LaneGraphEdgeMapObject], List[float]]:
-        """
-        Returns on-route lanes and heading errors where ego-vehicle intersects.
-        :param ego_state: state of ego-vehicle
-        :return: tuple of lists with lane objects and heading errors [rad].
-        """
-
-        ego_position_array: npt.NDArray[np.float64] = ego_state.rear_axle.array
-        ego_rear_axle_point: Point = Point(*ego_position_array)
-        ego_heading: float = ego_state.rear_axle.heading
-
-        intersecting_lanes = self._observation.drivable_area_map.intersects(ego_rear_axle_point)
-
-        on_route_lanes, on_route_heading_errors = [], []
-        for lane_id in intersecting_lanes:
-            if lane_id in self._route_lane_dict.keys():
-                # collect baseline path as array
-                lane_object = self._route_lane_dict[lane_id]
-                lane_discrete_path: List[StateSE2] = lane_object.baseline_path.discrete_path
-                lane_state_se2_array = np.array(
-                    [state.array for state in lane_discrete_path], dtype=np.float64
-                )
-                # calculate nearest state on baseline
-                lane_distances = (ego_position_array[None, ...] - lane_state_se2_array) ** 2
-                lane_distances = lane_distances.sum(axis=-1) ** 0.5
-
-                # calculate heading error
-                heading_error = lane_discrete_path[np.argmin(lane_distances)].heading - ego_heading
-                heading_error = np.abs(normalize_angle(heading_error))
-
-                # add lane to candidates
-                on_route_lanes.append(lane_object)
-                on_route_heading_errors.append(heading_error)
-
-        return on_route_lanes, on_route_heading_errors
