+++ nuplan_garage/planning/script/config/simulation/planner/pdm_closed_planner.yaml
@@ -1,0 +1,20 @@
+pdm_closed_planner:
+  _target_: nuplan_garage.planning.simulation.planner.pdm_planner.pdm_closed_planner.PDMClosedPlanner
+  _convert_: 'all'
+  
+  idm_policies:
+    _target_: nuplan_garage.planning.simulation.planner.pdm_planner.proposal.batch_idm_policy.BatchIDMPolicy
+    _convert_: 'all'
+    speed_limit_fraction: [0.2,0.4,0.6,0.8,1.0] # Target velocity as fractions of current speed-limit
+    fallback_target_velocity: 15.0 # Desired fallback velocity in free traffic [m/s] 
+    min_gap_to_lead_agent: 1.0 # Minimum relative distance to lead vehicle [m]
+    headway_time: 1.5 # Desired time headway. The minimum possible time to the vehicle in front [s]
+    accel_max: 1.5 # Maximum acceleration [m/s^2]
+    decel_max: 3.0 # Maximum deceleration (positive value) [m/s^2]
+    
+  lateral_offsets: [-1.0, 1.0] # Signed offsets from centerline (or null) [m]
+  trajectory_samples:  80 # Number of samples in output trajectory
+  proposal_samples: 40 # Number of samples in proposals 
+  sample_interval: 0.1 # Interval of trajectory/proposal samples [s]  
+  map_radius: 50 # Radius to consider around ego [m]
+
