+++ nuplan_garage/planning/simulation/planner/pdm_planner/utils/pdm_geometry_utils.py
