+++ requirements.txt
@@ -1,0 +1,1 @@
+nuplan-devkit==1.2.2
